requirements.txt
boto3==1.21.33
botocore==1.24.33
charset-normalizer==2.0.9
distlib==0.3.4
Django==3.2.9
django-cors-headers==3.10.1
django-redis==4.12.1
djangorestframework==3.12.4
elasticsearch8==8.6.2
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
importlib-metadata==4.11.3
packaging==21.3
pre-commit==2.17.0
python-dotenv==0.19.2
redis==4.2.2
requests==2.26.0
s3transfer==0.5.2
sentry-sdk==1.5.8
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.14.0
Werkzeug==2.0.2
python-slugify==6.1.2
PyMySQL==1.0.2
openpyxl==3.0.9
pycryptodome==3.17.0
openai==0.27.2
ndjson==0.3.1
python-dotenv